Fixes warning: implicit conversion changes signedness: 'NSInteger' (aka 'int') to 'NSUInteger' (aka 'unsigned int') when built with -Wsign-conversion

+++ AFNetworking/AFHTTPRequestOperation.m
@@ -387,7 +387,7 @@ didReceiveResponse:(NSURLResponse *)response
 totalBytesExpectedToWrite:(NSInteger)totalBytesExpectedToWrite
 {
     if (self.uploadProgress) {
-        self.uploadProgress(totalBytesWritten, totalBytesExpectedToWrite);
+        self.uploadProgress((NSUInteger)totalBytesWritten, (NSUInteger)totalBytesExpectedToWrite);
     }
 }
